IB/core: fix tautological compare of u16 with PAGE_SIZE
bundle_size (u16) and hdr.length (__u16) can never exceed PAGE_SIZE
on architectures with 64K pages (like powerpc).  Cast to u32 to make
the bounds check meaningful on all architectures.

+++ drivers/infiniband/core/uverbs_ioctl.c
@@ -55,7 +55,7 @@ void uapi_compute_bundle_size(struct uverbs_api_ioctl_method *method_elm,
 		ALIGN(bundle_size + 256, sizeof(*pbundle->internal_buffer));
 
 	/* Do not want order-2 allocations for this. */
-	WARN_ON_ONCE(method_elm->bundle_size > PAGE_SIZE);
+	WARN_ON_ONCE((u32)method_elm->bundle_size > PAGE_SIZE);
 }
 
 static bool uverbs_is_attr_cleared(const struct ib_uverbs_attr *uattr,
@@ -544,7 +544,7 @@ long ib_uverbs_ioctl(struct file *filp, unsigned int cmd, unsigned long arg)
 	if (err)
 		return -EFAULT;
 
-	if (hdr.length > PAGE_SIZE ||
+	if ((u32)hdr.length > PAGE_SIZE ||
 	    hdr.length != struct_size(&hdr, attrs, hdr.num_attrs))
 		return -EINVAL;
